# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplicationAnswers": {
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionAnswer"
            },
            "title": "Questions",
            "type": "array"
          }
        },
        "title": "ApplicationAnswers",
        "type": "object"
      },
      "ApplicationDetail": {
        "additionalProperties": true,
        "properties": {
          "application_answers": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationAnswers"
              },
              {
                "type": "null"
              }
            ]
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "candidate_profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Profile Id"
          },
          "cover_letter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Letter"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "jobs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobRecord"
              },
              {
                "type": "null"
              }
            ],
            "description": "Offre associée (jointure Supabase)"
          },
          "match_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "notion_page_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notion Page Id"
          },
          "prepared_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prepared At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tailored_cv": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TailoredCV"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "job_id",
          "status"
        ],
        "title": "ApplicationDetail",
        "type": "object"
      },
      "ApplicationIdResponse": {
        "properties": {
          "application_id": {
            "title": "Application Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message",
          "application_id"
        ],
        "title": "ApplicationIdResponse",
        "type": "object"
      },
      "ApplicationListResponse": {
        "properties": {
          "applications": {
            "items": {
              "$ref": "#/components/schemas/ApplicationDetail"
            },
            "title": "Applications",
            "type": "array"
          },
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "has_prev": {
            "title": "Has Prev",
            "type": "boolean"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "per_page": {
            "title": "Per Page",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "page",
          "per_page",
          "total_pages",
          "has_prev",
          "has_next",
          "applications"
        ],
        "title": "ApplicationListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobListResponse": {
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "has_prev": {
            "title": "Has Prev",
            "type": "boolean"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobRecord"
            },
            "title": "Jobs",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "per_page": {
            "title": "Per Page",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "page",
          "per_page",
          "total_pages",
          "has_prev",
          "has_next",
          "jobs"
        ],
        "title": "JobListResponse",
        "type": "object"
      },
...
      "NotionSyncResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "notion_page_id": {
            "title": "Notion Page Id",
            "type": "string"
          }
        },
        "required": [
          "message",
          "notion_page_id"
        ],
        "title": "NotionSyncResponse",
        "type": "object"
      },
...
      "UpdateStatusRequest": {
        "properties": {
          "status": {
            "enum": [
              "QUALIFIED",
              "PREPARING",
              "PREPARED",
              "READY",
              "APPLIED",
              "INTERVIEW",
              "INTERVIEW_HR",
              "INTERVIEW_TECH",
              "OFFER",
              "REJECTED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "UpdateStatusRequest",
        "type": "object"
      },
      "UpdateStatusResponse": {
        "properties": {
          "applied_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "notion_synced": {
            "default": false,
            "description": "Vrai si la fiche Notion liée a été mise à jour dans la foulée",
            "title": "Notion Synced",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "QUALIFIED",
              "PREPARING",
              "PREPARED",
              "READY",
              "APPLIED",
              "INTERVIEW",
              "INTERVIEW_HR",
              "INTERVIEW_TECH",
              "OFFER",
              "REJECTED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "message",
          "status"
        ],
        "title": "UpdateStatusResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend API moderne pour l'agent IA de recherche d'emploi et de génération de candidatures sur mesure.",
    "title": "Trema Job Search API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/applications": {
      "get": {
        "description": "Liste les candidatures existantes avec pagination et filtres.",
        "operationId": "list_applications_api_v1_applications_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Per Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "min_score",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Score"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Applications",
        "tags": [
          "Applications"
        ]
      }
    },
    "/api/v1/applications/create-from-job/{job_id}": {
      "post": {
        "description": "Crée une candidature à partir d'une offre d'emploi.",
        "operationId": "create_application_from_job_api_v1_applications_create_from_job__job_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationIdResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Application From Job",
        "tags": [
          "Applications"
        ]
      }
    },
    "/api/v1/applications/{app_id}": {
      "get": {
        "description": "Détail complet d'une candidature avec son offre associée.",
        "operationId": "get_application_api_v1_applications__app_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "App Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Application",
        "tags": [
          "Applications"
        ]
      }
    },
    "/api/v1/applications/{app_id}/documents/{doc_type}": {
      "get": {
        "description": "Prévisualise (inline) ou télécharge (`?download=true`) un document PDF : CV ou COVER_LETTER.",
        "operationId": "get_application_document_api_v1_applications__app_id__documents__doc_type__get",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "doc_type",
            "required": true,
            "schema": {
              "enum": [
                "CV",
                "COVER_LETTER"
              ],
              "title": "Doc Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "download",
            "required": false,
            "schema": {
              "default": false,
              "title": "Download",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {}
            },
            "description": "Document PDF généré (ReportLab)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Application Document",
        "tags": [
          "Applications"
        ]
      }
    },
    "/api/v1/applications/{app_id}/prepare": {
      "post": {
        "description": "Génère le CV personnalisé, la lettre et les réponses, puis synchronise Notion.",
        "operationId": "prepare_application_api_v1_applications__app_id__prepare_post",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "App Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationIdResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prepare Application",
        "tags": [
          "Applications"
        ]
      }
    },
    "/api/v1/applications/{app_id}/status": {
      "patch": {
        "description": "Met à jour le statut d'une candidature.",
        "operationId": "update_application_status_api_v1_applications__app_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "App Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Application Status",
        "tags": [
          "Applications"
        ]
      }
    },
    "/api/v1/applications/{app_id}/sync-notion": {
      "post": {
        "description": "Synchronise manuellement une candidature vers Notion.",
        "operationId": "sync_application_notion_api_v1_applications__app_id__sync_notion_post",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "App Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotionSyncResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Application Notion",
        "tags": [
          "Applications"
        ]
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "description": "Liste les offres enregistrées avec pagination et filtres.",
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Per Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "contract_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Contract Type"
            }
          },
          {
            "in": "query",
            "name": "min_score",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Score"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "Jobs"
        ]
      }
    },
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "description": "Détail complet d'une offre, avec la candidature associée si elle existe.",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "Jobs"
        ]
      }
    },
...
  }
}